add launch screen modifiy SceneDelegate, mainViewController
