+++ babPlus/babPlus/Controller/LaunchViewController.swift
@@ -1,0 +1,55 @@
+//
+//  LaunchViewController.swift
+//  babPlus
+//
+//  Created by Hongdonghyun on 2020/02/06.
+//  Copyright © 2020 YoujinMac. All rights reserved.
+//
+
+import UIKit
+
+class LaunchViewController: UIViewController {
+    let APPDELEGATE = UIApplication.shared.connectedScenes.first?.delegate as! SceneDelegate
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .red
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(true)
+        let mainVC = MainViewController()
+        let mainMapVC = BranchsAddressMapViewController()
+        let mainNaviController = UINavigationController(rootViewController: mainVC)
+        let mapNaviController = UINavigationController(rootViewController: mainMapVC)
+        mainNaviController.tabBarItem = UITabBarItem(title: "Home", image: UIImage(systemName: "house.fill"), tag: 0)
+        mapNaviController.tabBarItem = UITabBarItem(title: "Map", image: UIImage(systemName: "map.fill"), tag: 1)
+
+        let tabBarController = UITabBarController()
+        tabBarController.viewControllers = [mainNaviController, mapNaviController]
+        
+        RequestHelper().reqTask(path: "menu", method: "GET") {
+            (result) in
+            self.APPDELEGATE.dummy = result
+            DispatchQueue.main.async {
+                self.APPDELEGATE.window?.rootViewController = tabBarController
+            }
+            
+            
+        }
+        
+        
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
