+++ babPlus/babPlus/Controller/MainViewController.swift
@@ -21,6 +21,7 @@ class MainViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        let APPDELEGATE = UIApplication.shared.connectedScenes.first?.delegate as! SceneDelegate
         
         searchBarSet()
         setupCollectionView()
