+++ babPlus/babPlus/SceneDelegate.swift
@@ -10,62 +10,71 @@ import UIKit
 
 @available(iOS 13, *)
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    
     var window: UIWindow?
-
-
+    var dummy: Any?
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-
+        
         guard let windowScene = (scene as? UIWindowScene) else { return }
         
-        let mainVC = MainViewController()
-        let mainMapVC = BranchsAddressMapViewController()
+        let launchScreen = LaunchViewController()
         
-        let mainNaviController = UINavigationController(rootViewController: mainVC)
-        let mapNaviController = UINavigationController(rootViewController: mainMapVC)
         
-        mainNaviController.tabBarItem = UITabBarItem(title: "Home", image: UIImage(systemName: "house.fill"), tag: 0)
-        mapNaviController.tabBarItem = UITabBarItem(title: "Map", image: UIImage(systemName: "map.fill"), tag: 1)
-        
-        let tabBarController = UITabBarController()
-        tabBarController.viewControllers = [mainNaviController, mapNaviController]
+//        let mainVC = MainViewController()
+//        let mainMapVC = BranchsAddressMapViewController()
+//
+//        let mainNaviController = UINavigationController(rootViewController: mainVC)
+//        let mapNaviController = UINavigationController(rootViewController: mainMapVC)
+//
+//        mainNaviController.tabBarItem = UITabBarItem(title: "Home", image: UIImage(systemName: "house.fill"), tag: 0)
+//        mapNaviController.tabBarItem = UITabBarItem(title: "Map", image: UIImage(systemName: "map.fill"), tag: 1)
+//
+//        let tabBarController = UITabBarController()
+//        tabBarController.viewControllers = [mainNaviController, mapNaviController]
+//
+//        let window = UIWindow(windowScene: windowScene)
+//        window.rootViewController = tabBarController
+//        window.backgroundColor = .systemBackground
+//        window.makeKeyAndVisible()
+//        self.window = window
         
         let window = UIWindow(windowScene: windowScene)
-        window.rootViewController = tabBarController
+        window.rootViewController = launchScreen
         window.backgroundColor = .systemBackground
         window.makeKeyAndVisible()
         self.window = window
         
     }
-
+    
     func sceneDidDisconnect(_ scene: UIScene) {
         // Called as the scene is being released by the system.
         // This occurs shortly after the scene enters the background, or when its session is discarded.
         // Release any resources associated with this scene that can be re-created the next time the scene connects.
         // The scene may re-connect later, as its session was not neccessarily discarded (see `application:didDiscardSceneSessions` instead).
     }
-
+    
     func sceneDidBecomeActive(_ scene: UIScene) {
         // Called when the scene has moved from an inactive state to an active state.
         // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
     }
-
+    
     func sceneWillResignActive(_ scene: UIScene) {
         // Called when the scene will move from an active state to an inactive state.
         // This may occur due to temporary interruptions (ex. an incoming phone call).
     }
-
+    
     func sceneWillEnterForeground(_ scene: UIScene) {
         // Called as the scene transitions from the background to the foreground.
         // Use this method to undo the changes made on entering the background.
     }
-
+    
     func sceneDidEnterBackground(_ scene: UIScene) {
         // Called as the scene transitions from the foreground to the background.
         // Use this method to save data, release shared resources, and store enough scene-specific state information
         // to restore the scene back to its current state.
     }
-
-
+    
+    
 }
